chore: extracted out groupUnitsBySubjectCategory(...)
